Add a series of doublebond_intra groups specific to alkylaromatic rearrangements
These groups indicate that part of the chain
that cyclizes was originally part of a phenyl group.

This should help allow training reactions to find the right
group when training alkylaromatic cyclizations.

Either 1, 2, or 3 bonds are originally found in the
phenyl ring.  More than 3 bones is banned because
a shorter path can be found to the same transition state.